moved gen_exports to make testing easier
